+++ .github/workflows/ci.yml
@@ -26,10 +26,10 @@ jobs:
           - version: '1'  # The latest point-release (Windows)
             os: windows-latest
             arch: x64
-          - version: '1.6'  # 1.6 LTS (64-bit Linux)
+          - version: '1.10'  # 1.10 LTS (64-bit Linux)
             os: ubuntu-latest
             arch: x64
-          - version: '1.6'  # 1.6 LTS (32-bit Linux)
+          - version: '1.10'  # 1.10 LTS (32-bit Linux)
             os: ubuntu-latest
             arch: x86
     steps:
