+++ Project.toml
@@ -11,7 +11,7 @@ MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
 Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
 
 [extensions]
-PolyhedraExt = "Polyhedra"
+MultiObjectiveAlgorithmsPolyhedraExt = "Polyhedra"
 
 [compat]
 Combinatorics = "1"
@@ -19,8 +19,9 @@ HiGHS = "1"
 Ipopt = "1"
 JSON = "0.21"
 MathOptInterface = "1.19"
-Test = "<0.0.1, 1.6"
-julia = "1.6"
+Polyhedra = "0.8"
+Test = "1"
+julia = "1.10"
 
 [extras]
 HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
